Moved some functions to two other files.
